Add files via upload
Special areas of the Ludo board are typically coloured bright yellow, green, red, and blue. Each player is assigned a colour and has four tokens[c] in their colour. The board is normally square with a cross-shaped playspace, with each arm of the cross having three columns of squares, usually six per column. The middle columns usually have five squares coloured; these represent a player's home column. A sixth coloured square not on the home column is a player's starting square. At the centre of the board is a large finishing square, often composed of coloured triangles atop the players' home columns (thus depicting "arrows" pointing to the finish).